+++ lib/rivet-api-cloud/src/models/cdn_site_summary.rs
@@ -20,15 +20,15 @@ pub struct CdnSiteSummary {
     #[serde(rename = "display_name")]
     pub display_name: String,
     #[serde(rename = "create_ts")]
-    pub create_ts: i32,
+    pub create_ts: i64,
     #[serde(rename = "content_length")]
-    pub content_length: i32,
+    pub content_length: i64,
     #[serde(rename = "complete")]
     pub complete: bool,
 }
 
 impl CdnSiteSummary {
-    pub fn new(site_id: String, upload_id: String, display_name: String, create_ts: i32, content_length: i32, complete: bool) -> CdnSiteSummary {
+    pub fn new(site_id: String, upload_id: String, display_name: String, create_ts: i64, content_length: i64, complete: bool) -> CdnSiteSummary {
         CdnSiteSummary {
             site_id,
             upload_id,
