+++ lib/rivet-api-cloud/src/models/game_full.rs
@@ -16,7 +16,7 @@ pub struct GameFull {
     #[serde(rename = "game_id")]
     pub game_id: String,
     #[serde(rename = "create_ts")]
-    pub create_ts: i32,
+    pub create_ts: i64,
     #[serde(rename = "name_id")]
     pub name_id: String,
     #[serde(rename = "display_name")]
@@ -34,7 +34,7 @@ pub struct GameFull {
 }
 
 impl GameFull {
-    pub fn new(game_id: String, create_ts: i32, name_id: String, display_name: String, developer_team_id: String, total_player_count: i32, namespaces: Vec<crate::models::NamespaceSummary>, versions: Vec<crate::models::VersionSummary>, available_regions: Vec<crate::models::RegionSummary>) -> GameFull {
+    pub fn new(game_id: String, create_ts: i64, name_id: String, display_name: String, developer_team_id: String, total_player_count: i32, namespaces: Vec<crate::models::NamespaceSummary>, versions: Vec<crate::models::VersionSummary>, available_regions: Vec<crate::models::RegionSummary>) -> GameFull {
         GameFull {
             game_id,
             create_ts,
