+++ lib/rivet-api-cloud/src/models/namespace_summary.rs
@@ -16,7 +16,7 @@ pub struct NamespaceSummary {
     #[serde(rename = "namespace_id")]
     pub namespace_id: String,
     #[serde(rename = "create_ts")]
-    pub create_ts: i32,
+    pub create_ts: i64,
     #[serde(rename = "display_name")]
     pub display_name: String,
     #[serde(rename = "version_id")]
@@ -26,7 +26,7 @@ pub struct NamespaceSummary {
 }
 
 impl NamespaceSummary {
-    pub fn new(namespace_id: String, create_ts: i32, display_name: String, version_id: String, name_id: String) -> NamespaceSummary {
+    pub fn new(namespace_id: String, create_ts: i64, display_name: String, version_id: String, name_id: String) -> NamespaceSummary {
         NamespaceSummary {
             namespace_id,
             create_ts,
