+++ lib/rivet-api-cloud/src/models/version_full.rs
@@ -16,7 +16,7 @@ pub struct VersionFull {
     #[serde(rename = "version_id")]
     pub version_id: String,
     #[serde(rename = "create_ts")]
-    pub create_ts: i32,
+    pub create_ts: i64,
     #[serde(rename = "display_name")]
     pub display_name: String,
     #[serde(rename = "config")]
@@ -24,7 +24,7 @@ pub struct VersionFull {
 }
 
 impl VersionFull {
-    pub fn new(version_id: String, create_ts: i32, display_name: String, config: crate::models::CloudVersionConfig) -> VersionFull {
+    pub fn new(version_id: String, create_ts: i64, display_name: String, config: crate::models::CloudVersionConfig) -> VersionFull {
         VersionFull {
             version_id,
             create_ts,
