+++ lib/rivet-api-cloud/src/models/version_summary.rs
@@ -16,13 +16,13 @@ pub struct VersionSummary {
     #[serde(rename = "version_id")]
     pub version_id: String,
     #[serde(rename = "create_ts")]
-    pub create_ts: i32,
+    pub create_ts: i64,
     #[serde(rename = "display_name")]
     pub display_name: String,
 }
 
 impl VersionSummary {
-    pub fn new(version_id: String, create_ts: i32, display_name: String) -> VersionSummary {
+    pub fn new(version_id: String, create_ts: i64, display_name: String) -> VersionSummary {
         VersionSummary {
             version_id,
             create_ts,
